fix: split navigation provider exports
refs #737
